Fixed Bugs in Editing the Flashcard Set and also fixing the Next and Preview button outputs
