+++ src/pages/FlashcardSetModule/EditFlashcards/EditFlashcards.tsx
@@ -21,11 +21,11 @@ import {
 
 import { DeleteIcon } from "@chakra-ui/icons";
 import { EditFlashcardSetProps } from "../../../shared/types/typesEditFlashcardSet";
-import { useAddRemoveFlashcard } from "../../../shared/hooks/EditFlashcardSet/useAddRemoveFlashcard";
 import { useEditFlashcardSet } from "../../../shared/hooks/EditFlashcardSet/useEditFlashcardSet";
 import { useSaveEditedSet } from "../../../shared/hooks/EditFlashcardSet/useSaveEditedSet";
 import ButtonPrimary from "../../../shared/ui/components/ButtonPrimary";
 import { useValidation } from "../../../shared/hooks/EditFlashcardSet/useValidation";
+import { useAddRemoveFlashcard } from "../../../shared/hooks/EditFlashcardSet/useAddRemoveFlashcard";
 
 export default function EditFlashcards({
   isOpen,
